requirements.txt
pycountry
numpy
geopy
collection
streamlit
plotly.express
pandas
pymongo